+++ phenomena/connection_types/responder.py
@@ -5,7 +5,7 @@ from ontic.schema_type import SchemaType
 class Responder(OnticType):
     ONTIC_SCHEMA = SchemaType({
         'id': {
-            'type': int,
+            'type': long,
             'required': True,
         },
         'type': {
